running_time カラムを浮動小数点型に変更した

+++ database/migrations/2021_08_31_135020_create_user_quiz_results_table.php
@@ -19,7 +19,7 @@ class CreateUserQuizResultsTable extends Migration
             $table->integer('challenge_id'); 
             $table->integer('course_id'); 
             $table->integer('judgement')->nullable();
-            $table->integer('running_time')->nullable();
+            $table->float('running_time')->nullable();
             $table->timestamps();
             $table->unique(['user_id','course_id','challenge_id']);
         });
